+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## [Unreleased]
 
 ### Added
+
 - prevent claude from making changes when @ is on a session change (has Claude-session-id trailer)
+- split command
 
 ## [0.2.6] - 2025-10-13